mirgated main function to menu.cpp file
